requirements.txt
Flask>=2.0
python-dotenv>=0.19
google-generativeai>=0.3.0
# requests (if using REST APIs for flights/hotels etc.)
